backend/requirements.txt
apscheduler==3.9.1.post1
Flask==2.2.2
flask_cors==3.0.10
flask_restful==0.3.9
protobuf==4.21.12
pymongo==4.3.3
python-dotenv==0.21.0
